# feat: Enable TypeScript incremental compilation for 38-48% build performance improvement (Issue #37)