--- a/test-reports/junit-results.xml
+++ b/test-reports/junit-results.xml
@@ -1,2 +1,0 @@
-<testsuites id="" name="" tests="0" failures="0" skipped="0" errors="0" time="4.090049999999999">
-</testsuites>