… README.md
… package.json
… scratchpads/active/2025-08-24_stronger-success-detection-live-bookings-issue-30.md
… scratchpads/active/issue-37-typescript-incremental-builds-2025-09-01.md
… scratchpads/completed/issue-37-typescript-incremental-builds-2025-09-01.md
… src/utils/LiveDOMAnalyzer.ts
… src/utils/SuccessDetectionAnalytics.ts
… src/utils/SuccessDetectionRetryManager.ts
… src/utils/__tests__/SuccessDetectionAnalytics.test.ts
… test-artifacts/booking-flow-Squash-Bookin-3cab4-e-current-website-structure-chromium/test-finished-1.png
… test-artifacts/booking-flow-Squash-Bookin-3cab4-e-current-website-structure-chromium/video.webm
… test-artifacts/booking-flow-Squash-Bookin-745f4-ooking-flow-in-dry-run-mode-chromium/test-finished-1.png
… test-artifacts/booking-flow-Squash-Bookin-745f4-ooking-flow-in-dry-run-mode-chromium/video.webm
… test-artifacts/booking-flow-Squash-Bookin-bdda5-alculator-with-current-date-chromium-retry1/test-failed-1.png
… test-artifacts/booking-flow-Squash-Bookin-bdda5-alculator-with-current-date-chromium-retry1/trace.zip
… test-artifacts/booking-flow-Squash-Bookin-bdda5-alculator-with-current-date-chromium-retry1/video.webm
… test-artifacts/booking-flow-Squash-Bookin-bdda5-alculator-with-current-date-chromium/test-failed-1.png
… test-artifacts/booking-flow-Squash-Bookin-bdda5-alculator-with-current-date-chromium/trace.zip
… test-artifacts/booking-flow-Squash-Bookin-bdda5-alculator-with-current-date-chromium/video.webm
… test-artifacts/booking-flow-Squash-Bookin-bfafb-o-booking-page-successfully-chromium/test-finished-1.png
… test-artifacts/booking-flow-Squash-Bookin-bfafb-o-booking-page-successfully-chromium/video.webm
… test-reports/blob-report/report.zip
… test-reports/html-report/index.html
… test-reports/test-results.json
… tests/integration/SuccessDetectionIntegration.test.ts
… tests/unit/SuccessDetectionStrategies.test.ts
… tsconfig.json